Fixed VMM, (more) static ACPI support
Fixed a bug in which the PFA does not report the size of it’s state
table properly, leading to immediate corruption on writing to an
allocated storage descriptor in the address space manager. As well,
some files have been renamed to follow a standard convention, and the
static ACPI interface is now more complete. Additionally, a standard
kernel definition has been added for enabling debug printing on startup.