New PFA in progress because old is broken
Working on a split allocation page frame allocator because my current
one is not able to separate between it’s allocation users, as well,
something about the size of the kernel image seems to have broken it’s
somewhat fragile memory detection routines. This called for a complete
rewrite using and actual O(logN) physical allocator. The advantage of
this, other than speed, is that each memory region is individual to
it’s allocated tree, but free physical memory can be universally
coalesced. This allows for swap, even in the kernel, and generally,
more robustness and privatization of memory management. It also makes
it easier to add per-CPU caches for free pages, reducing global
contention on the physical allocator.