::MT::HW and ::HW conflict resolved.
Resolved a conflict in which using the ::HW namespace form inside
::MT::* would cause a namespace conflict due to ::MT::HW. Kernel
compiles again. Also, small addition, the AP’s are now set to wait for
the BSP before continuing, which may be useful for setup as well as
preventing premature startup problems from the INI-SIPI-SIPI protocol,
which may occur from time to time, as one SIPI is usually enough on
modern processors.