FADT
Wrote code to detect and use the FADT, which will allow for use of the
power management timer. This is important because it provides us with a
somewhat guaranteed alternative if the HPET isn’t available. (Note, the
TSC is still relied upon in the kernel where SMP may be operating, so
work must be done to fix that now and use either the PM Timer, or the
HPET if available. )