More AML work, swap in global GDT in APInit
Worked more on the AML parser and other dynamic ACPI stuff. Added in
the global GDT swap. Glad i caught that before implementing other mm
dependent things. It would suck if tasks intermittently seg-faulted
form a bad GDT! In theory, it isn’t necessary as the init image’s GDT
is currently identical to the kernel global GDT, but in practice, we
must leave room for improvement. This will keep userspace from crashing
instantly on being scheduled on an AP